+++ platform/ios_keyboard/ContainerApp/PrivatePinyinApp.swift
@@ -1,0 +1,10 @@
+import SwiftUI
+
+@main
+struct PrivatePinyinApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}